fixed harmless signed/unsigned comparison warning
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@36207 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/mac/carbon/notebmac.cpp
@@ -369,11 +369,12 @@ int wxNotebook::HitTest(const wxPoint& pt, long * flags) const
         }
     }
     
-    if ( outPart >= 1 && outPart <= countPages )
+    if ( outPart >= 1 && (size_t)outPart <= countPages )
     {
         resultV = outPart ;
     }    
-#endif
+#endif // TARGET_API_MAC_OSX
+
     if (flags != NULL)
     {
         *flags = 0;
